Fix broken movie links when the "About" section is hidden
Not sure what the problem is exactly but when the section was hidden there
were two issues:
- when hovering over a link the cursor doesn't change
to a pointer
- can't click on links (tabbing to a link and pressing Enter does work
however)

The problem has something to do with setting height and opacity to 0 to
hide the "About" section. Using "visibility: hidden" when the section is
hidden, can also be used as a fix, instead of setting the z-index.
I think that the z-index fix makes more sense because the movie link
list is supposed to be in front of the rest of the elements.